chore: 커스텀 에러 추가 및 이동
