+++ src/lib/error/index.ts
@@ -1,10 +1,2 @@
-export class CustomError extends Error {
-    errorType: string
-    i18nKey: string
-    
-    constructor(message: string, i18nKey: string) {
-        super(message)
-        this.errorType = 'CUSTOM_ERROR'
-        this.i18nKey = i18nKey
-    }
-}
+export * from './base'
+export * from './NetworkError'
